pick memories done, tremble might be broken
