+++ Videogame/LostIllusion/Assets/Scripts/map_scripts/Character_Movement.cs
@@ -108,6 +108,6 @@ public class Character_Movement : MonoBehaviour
         }
 
         // Reset position
-        transform.position = originalPosition;
+        //transform.position = originalPosition;
     }
 }
