+++ Videogame/LostIllusion/Assets/Scripts/map_scripts/pickMeUp.cs
@@ -12,32 +12,46 @@ using UnityEngine;
 public class pickMeUp : collideableObject
 {
     public bool interacted = false;
-    [SerializeField] public GameObject prefabDescriptionCanvas;
     [SerializeField] public GameObject prefabPickedObject;
     [SerializeField] public string cutsceneName;
+    private MapManager MapManager;
+
+    public override void Start()
+    {
+        base.Start();
+        MapManager = GameObject.Find("MapManager").GetComponent<MapManager>();
+        if (stateNameController.playedCutscenes.Contains(cutsceneName)){
+            Object.Instantiate(prefabPickedObject, transform.position, Quaternion.identity);
+            Destroy(gameObject);
+        }
+    }
     
     public override void onCollision(GameObject other)
     {
         base.onCollision(other);
         if (other.tag == "Player"){
             if (Input.GetKeyDown(KeyCode.Return)){
-                onPickUp();
+                onPickUp(other);
             }
         }
     }
 
-    public virtual void onPickUp()
+    public virtual void onPickUp(GameObject other)
     {
         if (!interacted){
         interacted = true;
-        GameObject pickedObject = Object.Instantiate(prefabPickedObject, transform.position, Quaternion.identity);
+        Object.Instantiate(prefabPickedObject, transform.position, Quaternion.identity);
 
         stateNameController.inventory.Add(gameObject.name);
         Debug.Log("Picked up: " + stateNameController.inventory[stateNameController.inventory.Count - 1]);
 
-        
-        
+        stateNameController.playerXPosInScene = other.transform.position.x;
+        stateNameController.playerYPosInScene = other.transform.position.y;
+        stateNameController.playerPreviousScene = UnityEngine.SceneManagement.SceneManager.GetActiveScene().name;
+
+        MapManager.StartCutscene(cutsceneName);
+        Destroy(gameObject);
+
         }
     }
-
 }
